Restructure database-related structures
