+++ consensus/src/model/stores/headers_selected_tip.rs
@@ -1,5 +1,5 @@
 use super::{
-    caching::{BatchDbWriter, CachedDbItem, DirectDbWriter},
+    database::prelude::{BatchDbWriter, CachedDbItem, DirectDbWriter},
     errors::StoreResult,
     DB,
 };
